Move DuplicatePageFor under Pages collection
